+++ PKHeX.Core/Saves/Substructures/Gen5/Entree/EntreeForest.cs
