+++ PKHeX.Core/Saves/Substructures/Gen5/Entree/EntreeForestArea.cs
