+++ PKHeX.Core/Saves/Substructures/Gen5/Entree/EntreeSlot.cs
