+++ PKHeX.Core/Saves/Substructures/Gen6/ConfigSave6.cs
