+++ PKHeX.Core/Saves/Substructures/Gen6/OPower/OPower6.cs
