+++ PKHeX.Core/Saves/Substructures/Gen6/OPower/OPower6Type.cs
