+++ PKHeX.Core/Saves/Substructures/Gen6/OPower/OPowerFlagSet.cs
