+++ PKHeX.Core/Saves/Substructures/Gen6/OPower/OPowerFlagState.cs
