+++ PKHeX.Core/Saves/Substructures/Gen7/LGPE/BelugaBlockIndex.cs
