+++ PKHeX.Core/Saves/Substructures/Gen7/LGPE/CaptureRecords.cs
