+++ PKHeX.Core/Saves/Substructures/Gen7/LGPE/ConfigSave7b.cs
