+++ PKHeX.Core/Saves/Substructures/Gen7/LGPE/EventWork7b.cs
