+++ PKHeX.Core/Saves/Substructures/Gen7/LGPE/GP1.cs
