+++ PKHeX.Core/Saves/Substructures/Gen7/LGPE/GoParkStorage.cs
