+++ PKHeX.Core/Saves/Substructures/Gen7/LGPE/IEventWork.cs
