+++ PKHeX.Core/Saves/Substructures/Gen7/LGPE/Misc7b.cs
