+++ PKHeX.Core/Saves/Substructures/Gen7/LGPE/MyItem7b.cs
