+++ PKHeX.Core/Saves/Substructures/Gen7/LGPE/MyStatus7b.cs
