+++ PKHeX.Core/Saves/Substructures/Gen7/LGPE/PlayTime7b.cs
