+++ PKHeX.Core/Saves/Substructures/Gen7/LGPE/PokeListHeader.cs
