+++ PKHeX.Core/Saves/Substructures/Gen7/LGPE/PouchSize7b.cs
