+++ PKHeX.Core/Saves/Substructures/Gen7/LGPE/WB7Records.cs
