+++ PKHeX.Core/Saves/Substructures/Gen7/LGPE/WR7GiftType.cs
