+++ PKHeX.Core/Saves/Substructures/Inventory/IItemFavorite.cs
@@ -1,0 +1,7 @@
+﻿namespace PKHeX.Core;
+
+public interface IItemFavorite
+{
+    /// <summary> Indicates if the item should be indicated as a favorite item. </summary>
+    bool IsFavorite { get; set; }
+}
