+++ PKHeX.Core/Saves/Substructures/Inventory/IItemNew.cs
@@ -1,0 +1,7 @@
+﻿namespace PKHeX.Core;
+
+public interface IItemNew
+{
+    /// <summary> Indicates if the item is "NEW"ly obtained and not yet viewed. </summary>
+    bool IsNew { get; set; }
+}
