+++ PKHeX.Core/Saves/Substructures/Inventory/Item/InventoryItem7.cs
