+++ PKHeX.Core/Saves/Substructures/Inventory/Item/InventoryItem7b.cs
