+++ PKHeX.Core/Saves/Substructures/Inventory/Item/InventoryItem8.cs
