+++ PKHeX.Core/Saves/Substructures/Inventory/Item/InventoryItem8a.cs
