+++ PKHeX.Core/Saves/Substructures/Inventory/Item/InventoryItem8b.cs
