+++ PKHeX.Core/Saves/Substructures/Inventory/MyItem.cs
