+++ PKHeX.Core/Saves/Substructures/Inventory/Pouch/InventoryPouch.cs
