+++ PKHeX.Core/Saves/Substructures/Inventory/Pouch/InventoryPouch3.cs
