+++ PKHeX.Core/Saves/Substructures/Inventory/Pouch/InventoryPouch3GC.cs
