+++ PKHeX.Core/Saves/Substructures/Inventory/Pouch/InventoryPouch4.cs
