+++ PKHeX.Core/Saves/Substructures/Inventory/Pouch/InventoryPouch7.cs
