+++ PKHeX.Core/Saves/Substructures/Inventory/Pouch/InventoryPouch7b.cs
