+++ PKHeX.Core/Saves/Substructures/Inventory/Pouch/InventoryPouch8.cs
