+++ PKHeX.Core/Saves/Substructures/Inventory/Pouch/InventoryPouch8a.cs
