+++ PKHeX.Core/Saves/Substructures/Inventory/Pouch/InventoryPouch8b.cs
