+++ PKHeX.Core/Saves/Substructures/Inventory/Pouch/InventoryPouchGB.cs
